feat: Integrate Keycloak into AuthContext
- Refactored user service into auth service
- Integrated keycloak profile into auth context
