+++ src/modules/Dashboard/containers/Dashboard/index.tsx
@@ -1,15 +1,17 @@
 import React from 'react';
 
-import UserService from '../../../../services/user_service';
+import { useAuth } from '../../../Auth/contexts/auth';
 import { Container, Header } from './styles';
 
 const Dashboard: React.FC = () => {
+  const { profile } = useAuth();
+
   return (
     <Container>
       <Header>
         <div>
           <h1>Dashboard</h1>
-          <span>Welcome back, {UserService.GetFirstName()}!</span>
+          <span>Welcome back, {profile?.firstName}!</span>
         </div>
       </Header>
     </Container>
