+++ src/routes/index.tsx
@@ -3,10 +3,9 @@ import React from 'react';
 import { Switch } from 'react-router-dom';
 
 import RouteWrapper, { Props as Route } from '../components/RouteWrapper';
-import Auth from '../modules/Auth/routes';
 import Dashboard from '../modules/Dashboard/routes';
 
-const routes: Route[] = [...Dashboard, ...Auth];
+const routes: Route[] = [...Dashboard];
 
 const Routes: React.FC = () => {
   return (
